Refactor JMS examples
