+++ spring-jms/src/main/java/com/baeldung/spring/jms/SampleJmsMessageSender.java
@@ -1,15 +1,10 @@
 package com.baeldung.spring.jms;
 
+import org.springframework.jms.core.JmsTemplate;
+
+import javax.jms.Queue;
 import java.util.HashMap;
 import java.util.Map;
-
-import javax.jms.JMSException;
-import javax.jms.Message;
-import javax.jms.Queue;
-import javax.jms.Session;
-
-import org.springframework.jms.core.JmsTemplate;
-import org.springframework.jms.core.MessageCreator;
 
 public class SampleJmsMessageSender {
 
@@ -25,16 +20,12 @@ public class SampleJmsMessageSender {
     }
 
     public void simpleSend() {
-        this.jmsTemplate.send(this.queue, new MessageCreator() {
-            public Message createMessage(Session session) throws JMSException {
-                return session.createTextMessage("hello queue world");
-            }
-        });
+        jmsTemplate.send(queue, s -> s.createTextMessage("hello queue world"));
     }
 
     public void sendMessage(final Employee employee) {
         System.out.println("Jms Message Sender : " + employee);
-        Map<String, Object> map = new HashMap<String, Object>();
+        Map<String, Object> map = new HashMap<>();
         map.put("name", employee.getName());
         map.put("age", employee.getAge());
         this.jmsTemplate.convertAndSend(map);
