+++ spring-jms/src/main/java/com/baeldung/spring/jms/SampleMessageConverter.java
@@ -1,12 +1,12 @@
 package com.baeldung.spring.jms;
 
+import org.springframework.jms.support.converter.MessageConversionException;
+import org.springframework.jms.support.converter.MessageConverter;
+
 import javax.jms.JMSException;
 import javax.jms.MapMessage;
 import javax.jms.Message;
 import javax.jms.Session;
-
-import org.springframework.jms.support.converter.MessageConversionException;
-import org.springframework.jms.support.converter.MessageConverter;
 
 public class SampleMessageConverter implements MessageConverter {
 
@@ -20,8 +20,7 @@ public class SampleMessageConverter implements MessageConverter {
 
     public Object fromMessage(Message message) throws JMSException, MessageConversionException {
         MapMessage mapMessage = (MapMessage) message;
-        Employee employee = new Employee(mapMessage.getString("name"), mapMessage.getInt("age"));
-        return employee;
+        return new Employee(mapMessage.getString("name"), mapMessage.getInt("age"));
     }
 
 }
